feat: move private components to be under one only
