--- src/stories/Go123Box.stories.jsx
@@ -1,14 +1,0 @@
-import { Go123Box } from "../components";
-
-export default {
-  title: "Go123 Box",
-  component: Go123Box,
-};
-
-export const Default = {
-  args: {
-    description: "Responding to",
-    bgColor: "bg-purple-200",
-    textColor: "text-black",
-  },
-};
