--- src/stories/SummaryBox.stories.jsx
@@ -1,15 +1,0 @@
-import SummaryBox from "../components/SummaryBox";
-
-export default {
-  title: "Summary Box",
-  component: SummaryBox,
-};
-
-export const Default = {
-  args: {
-    quantity: 1844,
-    quantifier: "participants",
-    bgColor: "bg-orange-500",
-    textColor: "text-white",
-  },
-};
